feat(config): disable federated sharing
With this config enabled users could enter user@host ID of external
Nextcloud instances. This should be disallowed for the MVP to avoid
unforseen consequences and to keep it simple.

The config "outgoing_server2server_share_enabled" was found by reverse
engineering the user lookup in the share UI.

The other configs were found by systematic search in the app source:

   grep -RE 'getApp.*files_sharing' apps/federatedfilesharing/

Signed-off-by: Thomas Lehmann <t.lehmann@strato.de>